Move User related code serializers and tests to users app.
